requirements.txt
APScheduler==3.9.1
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
docopt==0.6.2
Flask==2.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paramiko==2.11.0
prometheus-client==0.14.1
pycparser==2.21
PyNaCl==1.5.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uWSGI==2.0.20
Werkzeug==2.1.2
yarg==0.1.9
